src/requirements.txt
ui_tars==0.1.4
numpy==1.26.0
torch==2.7.0
moviepy==1.0.3
f5_tts==1.1.6
whisper==1.1.10
asyncio==3.4.3
Pillow==10.3.0
playwright==1.51.0
mcp==1.10.1
pydantic==2.10.6
camel-ai>=0.2.0
torchvision==0.22.0
PyMuPDF
whisperx
pdf2image
opencv_python